Added cards and chips to be able to select Physical Symptoms.
Using the same logic as I used for Allergens and Dietary Tags, chips are created for common physical symptoms that different food choices may help alleviate. The page now allows you to click common symptoms, describe how long they have been going on, and list how severe they are. The plan is based on this information to be able to make food suggestions that may help with the symptoms, and if the severity/length is concerning enough, recommend medical attention,.