Fixing a bug where you could not get back to the app from the  "Search Database", "Saved Foods", "Barcode Scan", and "Manual Entry" buttons.
Basically I just applied the same technique as I was already using for the "Add Photo" button to the rest of the buttons on the "Add Meal" page.

I just copied over barcode.tsx, saved.tsx, and search.tsx and moved them to the app\(tabs) folder so the dashboard at the bottom would still appear and users could escape these pop-ups once they are clicked.

In order to do this, I added 3 <Tabs.Screen> entries to the layout.tsx file.