Added Chalk and OpenAI to the package.json dependencies so app compiles for everybody upon npm install.
Before, when people pulled, they would have to manually install chalk to be able to run the servers for the app. Now, you can just run "npm install" for both the frontend and backend and it will work.

Same thing for OpenAI. This is for the API call. The app will run, but I am not going to commit my personal API key until I find a way to encrypt it, so the API call won't work unless I run it personally.