Completed the frontend for the PhotoAPI feature
Fairly simple, sends the image the user to OpenAI API (not right now because we don't have a true IP to work off of).

Tells the user the results of the API call if it works. If it doesn't it gives an error. I will add a loading spinner to this later.